+++ Areas/Public/Controllers/UsersTimelineController.cs
@@ -1,4 +1,4 @@
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Wayfarer.Models;
 using Wayfarer.Models.Dtos;
@@ -173,9 +173,7 @@ namespace Wayfarer.Areas.Public.Controllers
                     FullAddress = location.FullAddress,
                     ProviderAddressLine1 = location.ProviderAddressLine1,
                     AddressNumber = location.AddressNumber,
-                    ReverseGeocodingProvider = location.ReverseGeocodingProvider,
-                    ReverseGeocodingStorageMode = location.ReverseGeocodingStorageMode,
-                    ReverseGeocodedAt = location.ReverseGeocodedAt,
+                    IsGeoapifyAddress = location.IsGeoapifyAddress,
                     StreetName = location.StreetName,
                     PostCode = location.PostCode,
                     Place = location.Place,